+++ zarcel-utils/src/main/java/com/zing/zalo/adapter/PolymorphismZarcelAdapter.java
@@ -10,13 +10,11 @@ import com.zing.zalo.data.serialization.SerializedOutput;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.util.AbstractMap;
-import java.util.ArrayList;
-import java.util.Map;
+import java.util.*;
 
 public abstract class PolymorphismZarcelAdapter<T> implements ZarcelAdapter<T> {
 
-    private ArrayList<Map.Entry<Integer, Class>> children;
+    private TreeMap<Integer, Class> children;
 
     public enum RegisterType {
         ADD, UPDATE_CLASS, UPDATE_TYPE
@@ -25,69 +23,50 @@ public abstract class PolymorphismZarcelAdapter<T> implements ZarcelAdapter<T> {
     protected abstract void onRegisterChildClasses();
 
     protected PolymorphismZarcelAdapter() {
-        children = new ArrayList<>();
+        children = new TreeMap<>();
         onRegisterChildClasses();
     }
 
-    protected void register(int type, Class child, RegisterType registerType) throws ZarcelDuplicateException, ZarcelNotFoundException {
-        int index;
-        switch (registerType) {
-            case ADD:
-                index = haveType(type);
-                if (index != -1) {
-                    throw new ZarcelDuplicateException("Class " + child.getName() + " have already registered.");
-                } else {
-                    children.add(new AbstractMap.SimpleEntry<>(type, child));
-                }
-                break;
-            case UPDATE_TYPE:
-                index = haveType(type);
-                if (index == -1) {
-                    throw new ZarcelNotFoundException("Class " + child.getName() + " does not exist.");
-                } else {
-                    children.get(index).setValue(child);
-                }
-                break;
-            case UPDATE_CLASS:
-                index = haveClass(child);
-                if (index == -1) {
-                    throw new ZarcelNotFoundException("Class " + child.getName() + " does not exist.");
-                } else {
-                    children.remove(index);
-                    children.add(index, new AbstractMap.SimpleEntry<>(type, child));
-                }
-                break;
+    protected void registryAdd(int type, Class child) throws ZarcelDuplicateException {
+        if (children.containsKey(type)) {
+            throw new ZarcelDuplicateException("Class " + child.getName() + " have already registered.");
+        } else {
+            children.put(type, child);
         }
     }
 
-    private int haveType(int type) {
-        for (int index = 0; index < children.size(); index++) {
-            if (children.get(index).getKey() == type)
-                return index;
+    protected void registryUpdate(int type, Class child) throws ZarcelNotFoundException {
+        if (!children.containsKey(type)) {
+            throw new ZarcelNotFoundException("Class " + child.getName() + " does not exist.");
+        } else {
+            children.put(type, child);
         }
-        return -1;
     }
 
-    private int haveClass(Class child) {
-        for (int index = 0; index < children.size(); index++) {
-            if (children.get(index).getValue().equals(child))
-                return index;
+    protected void registryRemove(int type, Class child) throws ZarcelNotFoundException {
+        if (!children.containsKey(type)) {
+            throw new ZarcelNotFoundException("Class " + child.getName() + " does not exist.");
+        } else {
+            children.remove(type);
         }
-        return -1;
+    }
+
+    protected boolean isRegistered(int type) {
+        return children.containsKey(type);
     }
 
     @Override
     public void serialize(@NonNull T object, SerializedOutput writer) throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
         Method method;
-        for (int i = 0; i < children.size(); i++) {
-            if (children.get(i).getValue().isInstance(object)) {
-                writer.writeInt32(children.get(i).getKey());
-                method = children.get(i).getValue().getMethod("serialize", SerializedOutput.class);
+        for (Map.Entry<Integer, Class> child : children.entrySet()) {
+            if (child.getValue().isInstance(object)) {
+                writer.writeInt32(child.getKey());
+                method = child.getValue().getMethod("serialize", SerializedOutput.class);
                 method.invoke(object, writer);
                 return;
             }
         }
-        throw new ZarcelRuntimeException(object.getClass().getName() + "does not exist in Abstract Adapter.");
+        throw new ZarcelRuntimeException(object.getClass().getName() + "does not exist in Polymorphism Adapter.");
     }
 
     @Override
@@ -95,13 +74,13 @@ public abstract class PolymorphismZarcelAdapter<T> implements ZarcelAdapter<T> {
         Method method;
 
         int type = reader.readInt32();
-        for (int i = 0; i < children.size(); i++) {
-            if (children.get(i).getKey() == type) {
-                method = children.get(i).getValue().
+        for (Map.Entry<Integer, Class> child : children.entrySet()) {
+            if (child.getKey() == type) {
+                method = child.getValue().
                         getField("CREATOR")
                         .getType()
                         .getMethod("createFromSerialized", SerializedInput.class);
-                Object creator = children.get(i).getValue().newInstance().getClass().getDeclaredField("CREATOR");
+                Object creator = child.getValue().newInstance().getClass().getDeclaredField("CREATOR");
                 T childObject = (T) method.invoke(((Field) creator).get("CREATOR"), reader);
                 return childObject;
             }