+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+Zarcel v0.2.0 *(2018-08-13)*
+----------------------------
+* Thêm @Ignore để bỏ qua warning cho thuộc tính private.
+* Thay đổi signature của PolymorphismZarcelAdapter thành : registryAdd, registryUpdate, registryRemove.
+* Support @NotNull
+* Fix bugs.
+
 Zarcel v0.1.0 *(2018-08-10)*
 ----------------------------